--- app/src/main/java/io/horizontalsystems/bankwallet/modules/pin/main/PinContainerPresenter.kt
@@ -1,17 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.pin.main
-
-import androidx.lifecycle.ViewModel
-
-class PinContainerPresenter(
-        val router: PinContainerModule.Router,
-        private val showCancelButton: Boolean) : ViewModel(), PinContainerModule.ViewDelegate {
-
-    override fun onBackPressed() {
-        if (showCancelButton) {
-            router.closeActivity()
-        } else {
-            router.closeApplication()
-        }
-    }
-
-}
